requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
uvicorn>=0.22.0,<0.23.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
dotenv>=0.9.9